backend/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
idna==3.10
joblib==1.5.1
nltk==3.9.1
praw==7.8.1
prawcore==2.4.0
regex==2024.11.6
requests==2.32.4
textblob==0.19.0
tqdm==4.67.1
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.8.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 
